Use getContainedPoints() instead of getCoordinates()
The previously used method `Roi.getCoordinates()` is apparently doing something
different than returning the contents of the ROI, although even from reading the
documentation it is not clear what "define this selection" means.

Strangely enough, this has worked before and still is working with *most* ROIs
using the current sample dataset. Further testing revealed that using the
coordinates as pixel locations can be off by one (sometimes they are inside the
ROI and sometimes they are just one pixel outside), so this seems to be
describing the ROI's shape in some funny fashion.

Using the `Roi.getContainedPoints()` method now does the job properly.